requirements.txt
# Discord Bot
discord.py==2.3.2

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
itsdangerous==2.2.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Task Queue / Scheduler
apscheduler==3.10.4
redis==5.0.4
celery==5.4.0

# Configuration
pydantic==2.7.1
pydantic-settings==2.2.1

# Email
aiosmtplib==3.0.1
sendgrid==6.11.0

# Scraping
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.1

# Auth
authlib==1.3.0

# Discord Ed25519 signature verification
PyNaCl==1.5.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
